feat: code breakup for a scalable approach
