+++ src/UtilityFunctions.ts
@@ -1,0 +1,7 @@
+function sendMessage(webviewView: any, type: string, data: string) {
+  webviewView.webview.postMessage({ type: type, text: data });
+}
+
+export {
+    sendMessage,
+};
