+++ src/tools/ToolDefinations.ts
@@ -1,0 +1,5 @@
+import { read_defination } from "./ReadFile";
+
+export const toolDefinations = [
+    read_defination,
+];
