+++ src/types/common_types.ts
@@ -1,0 +1,11 @@
+export type Messages = Array<{
+  role: "user" | "system" | "assistant" | "tool";
+  content: string;
+  name?: string;
+}>;
+
+export const MessageType = {
+  TEXT: "text",
+  TOOL_USE: "tool_use",
+  ERROR: "error",
+};
